backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.3.5
spacy==3.7.2
scikit-learn==1.3.2
textacy==0.13.0
contractions==0.1.73
python-docx==1.1.0
docx2txt==0.8
plantuml==0.3.0
difflib2==0.1.1
aiofiles==0.24.0
